requirements.txt
apache-airflow
pandas
numpy
praw
configparser
s3fs
requests
beautifulsoup4